Corrected integration algorithm
Inserted input file input_truss_0 that is an simple example from Rao's book and verified differences between first integration method and central difference (corrected now).
The corrected control file now has call for both functions and a plot is generated to check similarities in behavior.
Damping coefficients leads to a completely different response when set to any value tha is not zero. Was observed that small values of coefficients leads to a great damping in first method. Check with newmark later and model in state space.